Turn system completed. Pathfinding fix
Turn systme is now completed. Hitting Enter will end the current
character's turn, begining the next character's turn. Also fixed a bug
where mousing over an inaccessible tile would still move the character
to that tile. There wouldn't be a path to it, but the character would
still move to it. Mousing over such a tile will return  an empty list
and the character will not move.